Allow build script to accept BUILD_VERSION argument. If no BUILD_VERSION argument is provided, use the local branch name
