+++ package.json
@@ -1,46 +1,46 @@
 {
-  "name": "gridappsd-viz",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "node server.js",
-    "launch": "webpack-dev-server --progress --inline",
-    "build": "rm -rf dist && node update-version.js && webpack"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@stomp/stompjs": "^4.0.2",
-    "bootstrap": "^4.1.3",
-    "d3": "4.7.4",
-    "express": "4.14.0",
-    "react": "^16.4.2",
-    "react-bootstrap": "^0.32.3",
-    "react-dom": "^16.4.2",
-    "react-json-viewer": "^1.7.1",
-    "react-router-dom": "^4.3.1",
-    "react-table": "^6.8.6",
-    "rxjs": "^6.2.2"
-  },
-  "devDependencies": {
-    "@types/d3": "4.7.0",
-    "@types/react": "^16.4.10",
-    "@types/react-dom": "^16.0.7",
-    "@types/react-table": "^6.7.21",
-    "@types/stompjs": "^2.3.2",
-    "awesome-typescript-loader": "3.4.1",
-    "css-loader": "0.28.11",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "1.1.11",
-    "html-webpack-plugin": "^3.2.0",
-    "jest": "^20.0.4",
-    "node-sass": "4.9.0",
-    "sass-loader": "7.0.1",
-    "source-map-loader": "0.1.5",
-    "style-loader": "0.21.0",
-    "typescript": "^3.0.1",
-    "webpack": "3.10.0",
-    "webpack-dev-server": "^2.9.7"
-  }
+    "name": "gridappsd-viz",
+    "version": "1.0.0",
+    "description": "",
+    "main": "index.js",
+    "scripts": {
+        "start": "node server.js",
+        "launch": "webpack-dev-server --progress --inline",
+        "build": "rm -rf dist && webpack --env=production"
+    },
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "@stomp/stompjs": "^4.0.2",
+        "bootstrap": "^4.1.3",
+        "d3": "4.7.4",
+        "express": "4.14.0",
+        "react": "^16.4.2",
+        "react-bootstrap": "^0.32.3",
+        "react-dom": "^16.4.2",
+        "react-json-viewer": "^1.7.1",
+        "react-router-dom": "^4.3.1",
+        "react-table": "^6.8.6",
+        "rxjs": "^6.2.2"
+    },
+    "devDependencies": {
+        "@types/d3": "4.7.0",
+        "@types/react": "^16.4.10",
+        "@types/react-dom": "^16.0.7",
+        "@types/react-table": "^6.7.21",
+        "@types/stompjs": "^2.3.2",
+        "awesome-typescript-loader": "3.4.1",
+        "css-loader": "0.28.11",
+        "extract-text-webpack-plugin": "^3.0.2",
+        "file-loader": "1.1.11",
+        "html-webpack-plugin": "^3.2.0",
+        "jest": "^20.0.4",
+        "node-sass": "4.9.0",
+        "sass-loader": "7.0.1",
+        "source-map-loader": "0.1.5",
+        "style-loader": "0.21.0",
+        "typescript": "^3.0.1",
+        "webpack": "3.10.0",
+        "webpack-dev-server": "^2.9.7"
+    }
 }
